refactor: add userFeedback component to handle errors and loaders
