+++ src/webapp/pages/analysis/steps/4-nursingMidwifery/NursingMidwiferyStep.tsx
@@ -5,7 +5,8 @@ import { useNursingMidwiferyStep } from "./useNursingMidwiferyStep";
 import { SelectMultiCheckboxes } from "$/webapp/components/selectmulti-checkboxes/SelectMultiCheckboxes";
 import { StepAnalysis } from "../StepAnalysis";
 import { PageStepProps } from "../../AnalysisPage";
-import { useLoading, useSnackbar } from "@eyeseetea/d2-ui-components";
+
+import { UserFeedbackContainer } from "$/webapp/components/user-feedback-container/UserFeedbackContainer";
 
 export const NursingMidwiferyStep: React.FC<PageStepProps> = React.memo(props => {
     const { analysis, section, title, updateAnalysis } = props;
@@ -18,32 +19,23 @@ export const NursingMidwiferyStep: React.FC<PageStepProps> = React.memo(props =>
         isLoading,
         error,
     } = useNursingMidwiferyStep({ analysis, section, updateAnalysis });
-    const loading = useLoading();
-    const snackbar = useSnackbar();
-
-    React.useEffect(() => {
-        if (isLoading) loading.show(isLoading, i18n.t("Running analysis..."));
-        else loading.hide();
-    }, [isLoading, loading]);
-
-    React.useEffect(() => {
-        if (error) snackbar.error(error);
-    }, [error, snackbar]);
 
     return (
-        <StepAnalysis
-            id={analysis.id}
-            section={section}
-            reload={reload}
-            title={title}
-            onRun={runAnalysis}
-        >
-            <SelectMultiCheckboxes
-                options={disaggregations}
-                onChange={handleChange}
-                value={selectedDisaggregations}
-                label={i18n.t("CatCombos")}
-            />
-        </StepAnalysis>
+        <UserFeedbackContainer isLoading={isLoading} error={error}>
+            <StepAnalysis
+                id={analysis.id}
+                section={section}
+                reload={reload}
+                title={title}
+                onRun={runAnalysis}
+            >
+                <SelectMultiCheckboxes
+                    options={disaggregations}
+                    onChange={handleChange}
+                    value={selectedDisaggregations}
+                    label={i18n.t("CatCombos")}
+                />
+            </StepAnalysis>
+        </UserFeedbackContainer>
     );
 });
